Adjustments from testing 0.3.0.
