+++ src/main/java/com/crunchydata/models/ColumnMetadata.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2024 the original author or authors.
+ * Copyright 2012-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-
-package com.crunchydata.model;
+package com.crunchydata.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -26,19 +25,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class DCTableMap {
-    private Integer tid;
-    private String destType = "target";
-    private String schemaName;
-    private String tableName;
-    private String modColumn;
-    private Integer parallelDegree = 1;
-    private String tableFilter;
-    private boolean tablePreserveCase = false;
-    private boolean schemaPreserveCase = false;
-    //Not from Table
-    private Integer batchNbr;
-    private String compareSQL;
-    private String tableAlias;
-    private Integer pid;
+public class ColumnMetadata {
+    public String columnList;
+    public Integer nbrColumns;
+    public Integer nbrPKColumns;
+    public String column;
+    public String pk;
+    public String pkList;
+    public String pkJSON;
 }
