+++ src/main/java/com/crunchydata/services/threadReconcile.java
@@ -24,10 +24,10 @@ import java.text.DecimalFormat;
 import java.util.concurrent.BlockingQueue;
 
 import com.crunchydata.controller.RepoController;
-import com.crunchydata.model.ColumnMetadata;
-import com.crunchydata.model.DCTable;
-import com.crunchydata.model.DCTableMap;
-import com.crunchydata.model.DataCompare;
+import com.crunchydata.models.ColumnMetadata;
+import com.crunchydata.models.DCTable;
+import com.crunchydata.models.DCTableMap;
+import com.crunchydata.models.DataCompare;
 import com.crunchydata.util.*;
 
 import static com.crunchydata.util.HashUtility.getMd5;
