+++ src/main/resources/pgcompare
@@ -1,0 +1,1 @@
+java -jar pgcompare.jar $@
