+++ src/main/resources/pgcompare.bat
@@ -1,0 +1,1 @@
+java -jar pgcompare.jar %*
